ethtool: Fix SFF-8472 transceiver module identification
According to table 5-1 in SFF-8472 Rev. 12.4, the Identifier Values for
"GBIC" (01h) and "Module soldered to motherboard (ex: SFF)" (02h) are
supported by the specification in addition to the current one.
Therefore, adjust ethtool to invoke the SFF-8079 parser for them, which
will in turn invoke the SFF-8472 parser if the transceiver module
supports digital diagnostic monitoring.

Without this patch, the EEPROM contents of such transceiver modules will
be hex dumped instead of being parsed and printed in a human readable
format.

Fixes: 25b64c66f58d ("ethtool: Add netlink handler for getmodule (-m)")
Reported-by: Ivar Simensen <is@datarespons.no>
Closes: https://lore.kernel.org/netdev/AM0PR03MB5938EE1722EF2C75112B86F5B9B9A@AM0PR03MB5938.eurprd03.prod.outlook.com/
Tested-by: Ivar Simensen <is@datarespons.no>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>